--- backend/filesystem/minimal/config.json
@@ -1,1 +1,0 @@
-{"name": "minimal", "config": {"editorShowAccentColors": true, "theme": "dark", "btOrder": "bottom-to-top"}}
