--- backend/filesystem/test_comp/code/graph.json
@@ -1,161 +1,0 @@
-{
-    "id": "fa2362dc-cffa-4764-9b70-90d475be0c02",
-    "offsetX": -112,
-    "offsetY": -35,
-    "zoom": 100,
-    "gridSize": 0,
-    "layers": [
-        {
-            "id": "1d9af8b5-858e-4253-9057-fbf3c9f2442d",
-            "type": "diagram-links",
-            "isSvg": true,
-            "transformed": true,
-            "models": {
-                "6bbaed1e-073a-4128-abc9-6208db8ae39f": {
-                    "id": "6bbaed1e-073a-4128-abc9-6208db8ae39f",
-                    "type": "default",
-                    "selected": false,
-                    "source": "94891f48-9deb-4724-a39d-3551b000dcf6",
-                    "sourcePort": "d988236c-73d4-423a-8acc-05911daede71",
-                    "target": "54aa79ff-3890-46d6-851f-1adcf80b7fb1",
-                    "targetPort": "45c3f019-7a89-4745-a5d4-91b9984650d3",
-                    "points": [
-                        {
-                            "id": "95145f71-f761-4c8d-98c9-f1f6816cead9",
-                            "type": "point",
-                            "x": 296.640625,
-                            "y": 221
-                        },
-                        {
-                            "id": "ab511423-0434-459d-b032-37dab54bbec0",
-                            "type": "point",
-                            "x": 377,
-                            "y": 265
-                        }
-                    ],
-                    "labels": [],
-                    "width": 3,
-                    "color": "gray",
-                    "curvyness": 50,
-                    "selectedColor": "rgb(0,192,255)"
-                },
-                "88fa7bc5-7a1a-49c6-b480-f4b05eed11da": {
-                    "id": "88fa7bc5-7a1a-49c6-b480-f4b05eed11da",
-                    "type": "default",
-                    "selected": false,
-                    "source": "94891f48-9deb-4724-a39d-3551b000dcf6",
-                    "sourcePort": "d988236c-73d4-423a-8acc-05911daede71",
-                    "target": "26bd9e94-9eaf-4313-970f-ebf18b05aa3c",
-                    "targetPort": "af680993-7e16-4890-8647-b56e5efe3c50",
-                    "points": [
-                        {
-                            "id": "bef04685-3ab7-4eee-910f-d914f749e3f0",
-                            "type": "point",
-                            "x": 296.640625,
-                            "y": 221
-                        },
-                        {
-                            "id": "31d8e48b-a9af-490e-b56d-ef6be4e0ff01",
-                            "type": "point",
-                            "x": 390,
-                            "y": 151
-                        }
-                    ],
-                    "labels": [],
-                    "width": 3,
-                    "color": "gray",
-                    "curvyness": 50,
-                    "selectedColor": "rgb(0,192,255)"
-                }
-            }
-        },
-        {
-            "id": "fed3b351-5f1a-4cad-9f61-031f8a6af57e",
-            "type": "diagram-nodes",
-            "isSvg": false,
-            "transformed": true,
-            "models": {
-                "94891f48-9deb-4724-a39d-3551b000dcf6": {
-                    "id": "94891f48-9deb-4724-a39d-3551b000dcf6",
-                    "type": "basic",
-                    "selected": false,
-                    "x": 200,
-                    "y": 200,
-                    "ports": [
-                        {
-                            "id": "d988236c-73d4-423a-8acc-05911daede71",
-                            "type": "children",
-                            "x": 291.640625,
-                            "y": 216,
-                            "name": "children",
-                            "alignment": "right",
-                            "parentNode": "94891f48-9deb-4724-a39d-3551b000dcf6",
-                            "links": [
-                                "6bbaed1e-073a-4128-abc9-6208db8ae39f",
-                                "88fa7bc5-7a1a-49c6-b480-f4b05eed11da"
-                            ],
-                            "in": false,
-                            "label": "children"
-                        }
-                    ],
-                    "name": "Tree Root",
-                    "color": "rgb(0,204,0)",
-                    "is_selected": false
-                },
-                "54aa79ff-3890-46d6-851f-1adcf80b7fb1": {
-                    "id": "54aa79ff-3890-46d6-851f-1adcf80b7fb1",
-                    "type": "basic",
-                    "selected": false,
-                    "x": 375,
-                    "y": 244,
-                    "ports": [
-                        {
-                            "id": "45c3f019-7a89-4745-a5d4-91b9984650d3",
-                            "type": "parent",
-                            "x": 372,
-                            "y": 260,
-                            "name": "parent",
-                            "alignment": "left",
-                            "parentNode": "54aa79ff-3890-46d6-851f-1adcf80b7fb1",
-                            "links": [
-                                "6bbaed1e-073a-4128-abc9-6208db8ae39f"
-                            ],
-                            "in": true,
-                            "label": "parent"
-                        }
-                    ],
-                    "name": "Test1",
-                    "color": "rgb(128,0,128)",
-                    "is_selected": false
-                },
-                "26bd9e94-9eaf-4313-970f-ebf18b05aa3c": {
-                    "id": "26bd9e94-9eaf-4313-970f-ebf18b05aa3c",
-                    "type": "basic",
-                    "selected": false,
-                    "x": 388,
-                    "y": 130,
-                    "ports": [
-                        {
-                            "id": "af680993-7e16-4890-8647-b56e5efe3c50",
-                            "type": "parent",
-                            "x": 385,
-                            "y": 146,
-                            "name": "parent",
-                            "alignment": "left",
-                            "parentNode": "26bd9e94-9eaf-4313-970f-ebf18b05aa3c",
-                            "links": [
-                                "88fa7bc5-7a1a-49c6-b480-f4b05eed11da"
-                            ],
-                            "in": true,
-                            "label": "parent"
-                        }
-                    ],
-                    "name": "Sub Tree",
-                    "color": "rgb(148,87,55)",
-                    "is_selected": false,
-                    "is_subtree": true
-                }
-            }
-        }
-    ]
-}
