--- backend/filesystem/test_comp/config.json
@@ -1,1 +1,0 @@
-{"name": "test_comp", "config": {"editorShowAccentColors": true, "theme": "dark", "btOrder": "bottom-to-top"}}
