+++ frontend/src/components/file_browser/FileBrowser.css
@@ -16,7 +16,6 @@
 
 .sidebar-entry-contents {
   overflow-y: auto;
-  flex: 1 1 1px;
 }
 
 .sidebar-entry-menu {
